Initial changes for ID update

+++ src/main/java/com/bloodmanagement/repository/DonorRepository.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.bloodmanagement.model.Donor;
 
 @Repository
-public interface DonorRepository extends JpaRepository<Donor, Long> {
+public interface DonorRepository extends JpaRepository<Donor, Integer> {
     Optional<Donor> findByPoliticalId(Integer politicalId);
 }
-
